auth fixes

+++ src/api/user/index.ts
@@ -1,7 +1,7 @@
 /* eslint-disable import/no-import-module-exports */
 /* eslint-disable consistent-return */
 import { Router } from 'express';
-import { encodeToJwt, validateApiKey } from '../../auth';
+import { encodeToJwt, verifyGoogle } from '../../auth';
 import logger from '../../logger';
 import {
   findByName, addUser, validateUserData, patchUser,
@@ -21,32 +21,7 @@ router.get('/:userName', async (req, res) => {
   }
   return res.json(user);
 });
-router.post('/login/:userName', async (req, res) => {
-  const { userName } = req.params;
-  const { apiKey } = req.body;
-  const isVerifiedToLogIn = validateApiKey(apiKey);
-  if (!isVerifiedToLogIn) {
-    return res.status(403).json('your api key is not verified for log in or no api key was provided.');
-  }
-  const userExists = await findByName(userName);
-  if (!userExists) {
-    res.status(404).json({ error: `User name ${userName} was not found.` });
-    return;
-  }
-  logger.info(`try login user by name ${userName}`);
-  try {
-    const token = encodeToJwt({ userName });
-    return res.json({
-      message: 'success login',
-      token,
-    });
-  } catch (e) {
-    logger.error(e);
-    res.status(500).json({
-      error: e,
-    });
-  }
-});
+
 router.post('/auth/:userName', async (req, res) => {
   const { userName } = req.params;
   logger.info(`try get user by name ${userName}`);
@@ -72,29 +47,36 @@ router.post('/auth/:userName', async (req, res) => {
     });
   }
 });
-router.get('/auth/validate/:userName', async (req, res) => {
+router.post('/login/:userName', async (req, res) => {
   try {
     const { userName } = req.params;
+    const { googleAuth } = req.body;
     const user = await findByName(userName);
     if (!user) {
       res.status(404).json({ error: `User name ${userName} was not found.` });
       return;
     }
-    const { code: userCodeInput } = req.query;
-    logger.info(`try validate user ${userName} entered correct pin code`);
-    await CacheService.init();
-    const correctCode = await CacheService.getKey(userName);
-    logger.info(`user code ${userCodeInput} correct code ${correctCode}`);
-    if (correctCode !== userCodeInput) {
-      logger.warn('wrong code');
-      return res.status(400).json({
-        error: 'user entered wrong pin code',
-      });
+    if (googleAuth) {
+      const googleVerified = await verifyGoogle(googleAuth, userName);
+      if (!googleVerified) {
+        return res.status(403).json('error verifying with google auth');
+      }
+    } else {
+      const { code: userCodeInput } = req.query;
+      logger.info(`try validate user ${userName} entered correct pin code`);
+      await CacheService.init();
+      const correctCode = await CacheService.getKey(userName);
+      logger.info(`user code ${userCodeInput} correct code ${correctCode}`);
+      if (correctCode !== userCodeInput) {
+        logger.warn('wrong code');
+        return res.status(400).json({
+          error: 'user entered wrong pin code',
+        });
+      }
+      CacheService.removeKey(userName);
+      logger.info('correct code');
     }
-    logger.info('correct code');
     const token = encodeToJwt({ userName });
-    // no need to await here
-    CacheService.removeKey(userName);
 
     return res.json({
       token,
